requirements.txt
streamlit
openai
pandas
python-docx
PyPDF2
python-dotenv
beautifulsoup4
requests


